requirements.txt
aiofiles==0.6.0
aiogram==2.11.2
aiohttp==3.7.3
appdirs==1.4.4
asgiref==3.3.4
async-timeout==3.0.1
attrs==19.3.0
autopep8==1.6.0
Babel==2.9.0
backcall==0.1.0
backports.entry-points-selectable==1.1.0
Beaker==1.11.0
beautifulsoup4==4.9.3
bleach==3.1.4
books==0.1.2
certifi==2021.5.30
chardet==3.0.4
charset-normalizer==2.0.7
click==7.1.2
colorama==0.4.3
cycler==0.10.0
decorator==4.4.2
defusedxml==0.6.0
distlib==0.3.2
Django==3.2.4
django-cors-headers==3.7.0
django-filter==2.4.0
django-js-asset==1.2.2
django-pandas==0.6.4
djangorestframework==3.12.4
djangorestframework-simplejwt==5.1.0
docutils==0.16
entrypoints==0.3
env==0.1.0
environs==9.3.2
fastapi==0.62.0
filelock==3.0.12
flake8==3.8.4
FormEncode==2.0.0
future==0.18.2
fuzzywuzzy==0.18.0
h11==0.11.0
html5lib==1.1
idna==3.2
ipykernel==5.2.0
ipython==7.13.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
jedi==0.16.0
Jinja2==2.11.2
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.2
jupyter-console==6.1.0
jupyter-core==4.6.3
kiwisolver==1.1.0
lockfile==0.12.2
Mako==1.1.3
MarkupSafe==1.1.1
marshmallow==3.12.1
matplotlib==3.2.1
mccabe==0.6.1
mistune==0.8.4
mpmath==1.1.0
multidict==5.1.0
mysqlclient==2.0.3
nbconvert==5.6.1
nbformat==5.0.5
nose==1.3.7
notebook==6.0.3
numpy==1.21.1
opencv-python==4.3.0.36
pandas==1.3.0
pandas-drf-tools==0.1.1
pandocfilters==1.4.2
parso==0.6.2
Paste==3.5.0
PasteDeploy==2.1.1
PasteScript==3.2.0
pickleshare==0.7.5
Pillow==8.2.0
pipenv==2021.5.29
platformdirs==2.0.2
ply==3.11
prometheus-client==0.7.1
prompt-toolkit==3.0.5
psycopg2==2.8.6
psycopg2-binary==2.9.1
pycodestyle==2.8.0
pydantic==1.7.3
pydruid==0.5.2
pyflakes==2.2.0
pygame==1.9.6
Pygments==2.7.3
PyJWT==2.3.0
Pylons==1.0.3
pyparsing==2.4.6
pyrsistent==0.16.0
pyspellchecker==0.5.4
pyTelegramBotAPI==3.6.7
pytesseract==0.3.5
python-daemon==2.2.4
python-dateutil==2.8.2
python-dotenv==0.18.0
python-Levenshtein==0.12.0
pytz==2021.1
pywin32==227
pywinpty==0.5.7
pyzmq==19.0.0
qtconsole==4.7.2
QtPy==1.9.0
repoze.lru==0.7
requests==2.26.0
rest-pandas==1.1.0
Routes==2.5.1
scipy==1.4.1
selenium==3.141.0
Send2Trash==1.5.0
simplejson==3.17.2
six==1.16.0
soupsieve==2.1
sqlparse==0.4.1
starlette==0.13.6
sympy==1.5.1
Tempita==0.5.2
terminado==0.8.3
testpath==0.4.4
toml==0.10.2
tornado==6.0.4
traitlets==4.3.3
typing-extensions==3.7.4.3
urllib3==1.26.7
uvicorn==0.13.0
virtualenv==20.4.7
virtualenv-clone==0.5.4
waitress==1.4.4
wcwidth==0.1.9
webencodings==0.5.1
WebError==0.13.1
WebHelpers==1.3
WebOb==1.8.6
WebTest==2.0.35
widgetsnbextension==3.5.1
XlsxWriter==3.0.2
yarl==1.6.3
